+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file, following t
 Note that since we don't clearly distinguish between a public and private interfaces there will be changes in non-major versions that are potentially breaking. If we make breaking changes to less used interfaces we will highlight it in here.
 
 ## [Unreleased]
+- TM-align performance improvements (#1745)
 - Disable transparent outline close to opaque elements
 - Add axis param to trackball spin & rock animation
 - Color smoothing fixes (#1747)